Py: Test cleanup
Move shared code to tests_common.py
Make ewald_gpu and p3m_gpu tests actually execute (wrong feature checks)
